modularize NLP code / improve error handling
